#13:  [Add public]

+++ app/src/main/java/com/extesion/ExtUtil.java
@@ -5,6 +5,6 @@ public class ExtUtil {
     public static float calDistance(float var0, float var1, float var2, float var3) {
         var0 -= var2;
         var1 -= var3;
-        return (float)Math.sqrt((double)(var0 * var0 + var1 * var1));
+        return (float)Math.sqrt(var0 * var0 + var1 * var1);
     }
 }
